sepehrsmartservices/Eghamat24/requirements.txt
django_environ==0.12.0
Flask==3.1.0
lxml==5.3.2
redis==5.2.1
Requests==2.32.3
gunicorn==23.0.0
